commit de emergência
Todas as atualizações necessárias pra essa branch em um commit.

+++ src/main/java/br/edu/ifto/ecommerce/model/entity/venda/ItemVenda.java
@@ -24,7 +24,7 @@ public class ItemVenda implements Serializable {
     @ManyToOne
     private Venda venda;
 
-    private Double quantidade;
+    private int quantidade;
 
     public BigDecimal total(){
         return this.produto
